Gates: Improve Exits - now modifiable, exit center
Exits are now modifiable; you can use a Waygate Constructor on an existing gate and if you are the owner or an admin you can modify the exit of the gate to where you are standing.

Similarly to the fix in Waypoints, Gates now teleport users to the center of a block and not the corner, helping to fix issues where the exit is space constrained.