Let’s not check player head meta information
Player heads don’t validate correctly even if they are the same head, and this causes problems in identifying the menu.

As most menu items are custom items with names and lore, identifying the correct menu can be done fine without the need of delving too deep into player head metadata.

So in this case, we can just verify the two items as player heads and don’t check equality on the heads themselves, relying on the other items to ensure a complete and trusted verification of the menu.